+++ NEWS.md
@@ -7,7 +7,7 @@ Cite all releases with doi [10.5281/zenodo.5118872](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
-# matsbyname 0.6.8 (2023-12-20)
+# matsbyname 0.6.8 (2023-12-20) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10418797.svg)](https://doi.org/10.5281/zenodo.10418797)
 
 * Added a statement of need.
 * Updated the GitHub pages approach to use the gh-pages branch.