+++ NEWS.md
@@ -7,6 +7,8 @@ Cite all releases with doi [10.5281/zenodo.5118872](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
+## matsbyname 0.6.9 (2024-01-30)
+
 * Now running tests in parallel.
 * Added performance tests.
 * Fixed a bug in `replaceNaN_byname()`